Refactor project structure and update build scripts
- Reorganize hardware modules:
  - Moved arc, blackhole, wormhole, grayskull, and quasar code into new subpackages under ttexalens/hw/tensix
  - Updated device creation logic in ttexalens/device.py to import from these new paths
- Reorganize GDB modules:
  - Moved gdb_communication, gdb_data, gdb_file_server, and gdb_server into ttexalens/gdb
  - Updated related imports (e.g. in ttexalens/uistate.py)
- Update NOC overlay generation:
  - Changed overlay outputs: wormhole to ttexalens/arch/wormhole/noc_overlay.py and blackhole to ttexalens/arch/blakchole/noc_overlay.py
- Adjust wheel build/install scripts:
  - Updated wheel install path in wheel_build_install.sh to build/ttexalens_wheel
- Extend setup.py:
  - Added a new "hw" configuration to copy only "*.py" files from ttexalens/hw

This refactoring improves code modularity and clarifies boundaries for hardware and debugging components.
