+++ scripts/wheel_build_install.sh
@@ -8,4 +8,4 @@ make build
 make wheel
 
 pip uninstall ttexalens -y
-pip install build/ttlens_wheel/*.whl
+pip install build/ttexalens_wheel/*.whl
