+++ ttexalens/gdb/gdb_data.py
