+++ ttexalens/hw/tensix/blackhole/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+
+# SPDX-License-Identifier: Apache-2.0
