+++ ttexalens/hw/tensix/grayskull/grayskull_ops.py
